refactor(nostr): extract wipeAccountCaches; watch arming logs viewer prefix (#279)
The identity-switch wiring pushed NostrContext.tsx past its file-size
baseline (1838 → 1872). wipeAccountCaches was a zero-closure useCallback
(empty deps) — now a pure module (src/contexts/accountCacheWipe.ts),
taking the file to 1785. Also: the [BgDmWatch] arming marker logs an
8-char viewer pubkey prefix (used to verify the switch re-arm on-device),
and the dedupe doc comment states the missing-id contract (Copilot nit).
